Minimise crate dependancy on graphql_types
