+++ server/graphql/cold_chain/src/mutations/temperature_breach.rs
@@ -1,9 +1,9 @@
+use crate::types::temperature_breach::TemperatureBreachNode;
 use async_graphql::*;
 use graphql_core::{
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
-use graphql_types::types::TemperatureBreachNode;
 use repository::TemperatureBreach;
 use service::{
     auth::{Resource, ResourceAccessRequest},
