+++ server/graphql/cold_chain/src/types/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod sensor;
+pub(crate) mod temperature_breach;
+pub(crate) mod temperature_log;
+pub(crate) mod temperature_notification;
