+++ server/graphql/general/src/lib.rs
@@ -5,14 +5,12 @@ mod sync_api_error;
 pub use self::queries::sync_status::*;
 use self::queries::*;
 
-use chrono::{DateTime, Utc};
 use graphql_core::pagination::PaginationInput;
 use util::is_central_server;
 
 use crate::store_preference::store_preferences;
 use graphql_types::types::{
     CurrenciesResponse, CurrencyFilterInput, CurrencySortInput, StorePreferenceNode,
-    TemperatureLogFilterInput,
 };
 use mutations::{
     barcode::{insert_barcode, BarcodeInput},
@@ -297,25 +295,6 @@ impl GeneralQueries {
 
     pub async fn plugins(&self, ctx: &Context<'_>) -> Result<Vec<PluginNode>> {
         get_plugins(ctx)
-    }
-
-    pub async fn temperature_chart(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        #[graphql(desc = "Must be before toDatetime")] from_datetime: DateTime<Utc>,
-        #[graphql(desc = "Must be after fromDatetime")] to_datetime: DateTime<Utc>,
-        #[graphql(desc = "Minimum 3 and maximum 100")] number_of_data_points: i32,
-        filter: Option<TemperatureLogFilterInput>,
-    ) -> Result<TemperatureChartResponse> {
-        temperature_chart(
-            ctx,
-            store_id,
-            from_datetime,
-            to_datetime,
-            number_of_data_points,
-            filter,
-        )
     }
 
     pub async fn currencies(
