+++ server/graphql/general/src/queries/mod.rs
@@ -47,8 +47,6 @@ pub use self::log::*;
 pub mod last_successful_user_sync;
 pub use self::last_successful_user_sync::*;
 pub use self::plugin::*;
-pub mod temperature_chart;
-pub use self::temperature_chart::*;
 pub mod currency;
 
 #[cfg(test)]
